CreatedSpriteMakerApp

+++ GameAPI/Player.cs
@@ -2,7 +2,7 @@
 {
     public class Player : GameObject
     {
-        public Player(float x, float y) : base(x, y)
+        public Player(int x, int y) : base(x, y)
         {
 
         }
